requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pypdf
langchain
langchain-community
langchain-openai
langgraph
langchain-chroma
chromadb
openai
tiktoken
faiss-cpu
